Differentiate js_str! and js_err! macros
js_err had nothing to do with an error, actually.
It returned js_sys::String which happened to be often used
in place of returning errors to JavaScript runtime.
js_str becomes the one that does that role, and js_err now
returns Err variant of std::result::Result, which can be actually
used with errors.