Move objects on the transport belt align to the center
It is not realistic, but make the flow much smoother.

Just like Factorio, transport belts should have abilities to align items on them,
otherwise it would be a mess, spilling items over after a few turns.

The difference between this game and Factorio is that Factorio has two lanes
on a belt, but it would complicate factory design too much.
Also, item transition movements on curves and merges are not optimized as
Factorio. We're using WebAssembly, there is a limit to optimize anyway.