Use StructureId to calculate fluid in O(n)
It has been O(n^2) since we checked every combination in the
entire game. Without some kind of spatial index, it would be
inevitable. However, with stable StructureId as the index into
the Vec, we can quickly get the reference to the structure.

The tricky part is that We need to exclude self from accessible
Structure list when calling a method, but we managed to implement
StructureDynIter that does split the Vec into halves and still look up
in constant time.