Use serde-wasm-bindgen to group parameters together
We could use serde_json as we do for save files, but it seems to
bloat the code size (although a bit, about 4KB).

serde-wasm-bindgen is a great way to pass parameters from JS
to Rust-wasm, because it wouldn't go through JSON serialization
and floating point precision loss in string formatting.