Maintain drop_items_index between frames to avoid reconstruction
We were rebuilding the DropItemIndex from scratch every frame,
but it isn't necessary. We could keep the cache from previous
frame and update only if an item crosses a border between chunks.

In theory, it would reduce the number of reallocations greatly, but
in practive building index is pretty fast compared to actual collision
checking, so it wouldn't change much.