Limit the size of chest storage inventory to 48 slots
Because Factorio's storage capacity is measured by slots,
it is not easy to determine if you can put some new items.

We could manage slots in flat data structure, but it would
cost more memory and computation time to determine
if an inventory contains some given item type.
HashMap is intuitive, fast and easy to implement as
an inventory.

I don't make final decision on the data structure, but it seems
to be working for now, so I will leave it for the time being.