Put items into structure inventories as much as possible
In move_all_inventory_items, we collect player's item count
into a Vec and try to move all items even if one of the item types
fails to fit to a structure. It yields a better behavior for factory
or burner, where the ingredients will be pushed to input
buffers.