Reduce maximum delta time to avoid frame accumulation
Even in fast computers, if you switch a browser tab, it will become
dormant and won't run requestAnimationFrame, or when a periodic game
save runs, it can make the frames accumulate.
It will run many frames without rendering like crazy to catch up with most recent
timer. We want to avoid it. Also, it's not much use if the computer is so slow
that you drop more than 0.2s * 60fps = 12 frames in redering, because it would
be too laggy to play.